+++ src/utils.h
@@ -2,7 +2,6 @@
 #define UTILS_H
 
 #include <stdlib.h>
-#include <stdint.h>
 
 char * chomp (char *str);
 char * trim  (char *str);
@@ -11,7 +10,7 @@ char * trimc (char *str, int c);
 size_t buf_expand (void **buf, size_t size,
 		size_t old_nmemb, size_t length);
 
-uint32_t powu (uint32_t base, uint32_t exp);
+size_t powu (size_t base, size_t exp);
 
 int exists      (const char *file);
 int is_terminal (int fd);
